Simplify compiler wrappers, following refactorings in AllocsCC. Change naming of ClangAllocsCC source file, for consistency.
