+++ bin/clangallocscc
@@ -1,1 +1,1 @@
-../lib/ClangAllocsCC.py
+../lib/clangallocscc.py
